+++ model/question.go
@@ -4,12 +4,12 @@ import "time"
 
 // 映射表 question
 type Question struct {
-	QuestionID  uint64            `xorm:"PK AUTOINCR 'question_id'"`
+	QuestionID  uint32            `xorm:"PK AUTOINCR 'question_id'"`
 	Description string            `xorm:"description"`
-	Type        uint64            `xorm:"type"`
+	Type        uint32            `xorm:"type"`
 	Options     map[string]string `xorm:"options"`
 	Answer      string            `xorm:"answer"`
-	Duration    uint64            `xorm:"duration"`
+	Duration    uint32            `xorm:"duration"`
 	UpdatedAt   time.Time         `xorm:"DateTime UPDATED 'updated_at'"`
 	CreatedAt   time.Time         `xorm:"DateTime CREATED 'created_at'"`
 }
